+++ src/Math/Vector/2D/NRE_Vector2D.tpp
@@ -81,13 +81,13 @@
             }
 
             template <class T>
-            inline double Vector2D<T>::norm() const {
+            inline float Vector2D<T>::norm() const {
                 return std::sqrt(normSquared());
             }
 
             template <class T>
-            inline double Vector2D<T>::normSquared() const {
-                return static_cast <double> (x * x + y * y);
+            inline float Vector2D<T>::normSquared() const {
+                return static_cast <float> (x * x + y * y);
             }
 
             template <class T>
@@ -153,7 +153,7 @@
             }
 
             template <class T>
-            inline double Vector2D<T>::operator |=(Vector2D const& u) const {
+            inline float Vector2D<T>::operator |=(Vector2D const& u) const {
                 return x * u.x + y * u.y;
             }
 
@@ -193,7 +193,7 @@
             }
 
             template <class T>
-            inline double Vector2D<T>::operator |(Vector2D const& u) const {
+            inline float Vector2D<T>::operator |(Vector2D const& u) const {
                 return *this |= u;
             }
 
@@ -205,13 +205,13 @@
 
             template <>
             template <class K>
-            inline bool Vector2D<double>::operator ==(Vector2D<K> const& u) const {
+            inline bool Vector2D<float>::operator ==(Vector2D<K> const& u) const {
                 return almostEqual(x, u.x) && almostEqual(y, u.y);
             }
 
             template <>
             template <class K>
-            inline bool Vector2D<double>::operator ==(Vector2D<K> const& u) const {
+            inline bool Vector2D<float>::operator ==(Vector2D<K> const& u) const {
                 return almostEqual(x, u.x) && almostEqual(y, u.y);
             }
