+++ src/Math/Vector/3D/NRE_Vector3D.tpp
@@ -118,13 +118,13 @@
             }
 
             template <class T>
-            inline double Vector3D<T>::norm() const {
+            inline float Vector3D<T>::norm() const {
                 return std::sqrt(normSquared());
             }
 
             template <class T>
-            inline double Vector3D<T>::normSquared() const {
-                return static_cast <double> (x * x + y * y + z * z);
+            inline float Vector3D<T>::normSquared() const {
+                return static_cast <float> (x * x + y * y + z * z);
             }
 
             template <class T>
@@ -196,7 +196,7 @@
             }
 
             template <class T>
-            inline double Vector3D<T>::operator |=(Vector3D const& u) const {
+            inline float Vector3D<T>::operator |=(Vector3D const& u) const {
                 return x * u.x + y * u.y + z * u.z;
             }
 
@@ -244,7 +244,7 @@
             }
 
             template <class T>
-            inline double Vector3D<T>::operator |(Vector3D const& u) const {
+            inline float Vector3D<T>::operator |(Vector3D const& u) const {
                 return *this |= u;
             }
 
@@ -261,13 +261,13 @@
 
             template <>
             template <class K>
-            inline bool Vector3D<double>::operator ==(Vector3D<K> const& u) const {
+            inline bool Vector3D<float>::operator ==(Vector3D<K> const& u) const {
                 return almostEqual(x, u.x) && almostEqual(y, u.y) && almostEqual(z, u.z);
             }
 
             template <>
             template <class K>
-            inline bool Vector3D<double>::operator ==(Vector3D<K> const& u) const {
+            inline bool Vector3D<float>::operator ==(Vector3D<K> const& u) const {
                 return almostEqual(x, u.x) && almostEqual(y, u.y) && almostEqual(z, u.z);
             }
